fw_iso_ctx: code refactoring to split allocate function
It's planned to make Hinoko.FwIsoCtx as interface. It's convenient
for derived object to use private helper function instead of instance
method.

This commit splits current implementaion of Hinoko.FwIsoCtx.allocate()
into internal API part and private implementation.
